requirements.txt
gunicorn
whitenoise
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
PyJWT==1.7.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.13.1
